Refactor move DeleteConversationDialog to 'gui.conversation' namespace
